added lambda_exprt
This introduces lambda_exprt, derived from binding_exprt.  This is Church's
lambda calculus, i.e., a nameless mathematical function, and is not to be
confused with lambda expressions in Java or C++.

The C front-end can parse these using ACSL syntax.  The simplifier is
extended to be able to simplify function applications of lambda expressions.